Fix APPUiO Cloud queries to gracefully handle control-api timeseries overlap
We've observed that there's a chance that we have timeseries from two
different control-api pods for a short period of time during the
maintenance.

This commit adjust the APPUiO Cloud queries to drop the pod and instance
labels on the `appuio_control_organization_info` timeseries to ensure
that we don't have duplicate timeseries which break the query.