Extend the VSHN Managed OpenShift queries to tolerate changes to unused labels
We extend the queries with `max by (relevant_label_set)` where necessary
to ensure that changes to unused labels (e.g. `cluster_name`) don't
break the billing queries.

Note that the list of labels which we keep for `node_cpu_info` must
contain `core` and `cpu` for the vCPU-Hour billed clusters. The same
list must contain `core` but mustn't contain `cpu` for the Core-Hour
billed clusters.